Refactored cutsom attribute to be in line with other library attributes.
