--- Attributes/ValidationAttributes/MaxFileSizeAttribute.cs
@@ -1,31 +1,0 @@
-﻿using Microsoft.AspNetCore.Http;
-using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace GovUkDesignSystem.Attributes.ValidationAttributes
-{
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property, AllowMultiple = false, Inherited = true)]
-    public class MaxFileSizeAttribute : ValidationAttribute
-    {
-        private readonly int _maxFileSize;
-        public MaxFileSizeAttribute(int maxFileSize)
-        {
-            _maxFileSize = maxFileSize;
-        }
-
-        public override bool IsValid(object value)
-        {
-            var file = value as IFormFile;
-            if (file == null)
-            {
-                return false;
-            }
-            return file.Length <= _maxFileSize;
-        }
-
-        public override string FormatErrorMessage(string name)
-        {
-            return base.FormatErrorMessage(_maxFileSize.ToString());
-        }
-    }
-}
